requirements.txt
Flask==3.0.0
python-pptx==0.6.23
groq==0.4.2
requests==2.31.0
python-dotenv==1.0.0
